api/requirements.txt
fastapi
requests
uvicorn[standard]
pymongo[srv]
websockets
PyJWT
python-decouple
pydantic[email]